chore: dedupe script retry + log helpers
