+++ tests/test_write_artifacts.py
@@ -16,6 +16,9 @@ REPO_ROOT = Path(__file__).resolve().parents[1]
 
 def load_script_module(module_name: str, relative_path: str) -> ModuleType:
     module_path = REPO_ROOT / relative_path
+    scripts_dir = str(module_path.parent)
+    if scripts_dir not in sys.path:
+        sys.path.insert(0, scripts_dir)
     spec = importlib.util.spec_from_file_location(module_name, module_path)
     if spec is None or spec.loader is None:
         raise RuntimeError(f"Unable to load module from {module_path}")
